hygiene, widen types before doing uint arith

+++ lib/dt-core-strpf.c
@@ -417,7 +417,7 @@ __strfdt_dur(
 			if (that.d.durtyp != DT_DURD) {
 				return 0U;
 			}
-			dv = that.d.dv * HOURS_PER_DAY;
+			dv = (int64_t)that.d.dv * HOURS_PER_DAY;
 			/*@fallthrough@*/
 		case DT_DURH:
 			dv *= MINS_PER_HOUR;
